Create CarCRUDTest and change the way cars are instantiated
