--- RentACar/RentACar/CRUD/Cars.json
@@ -1,18 +1,0 @@
-[
-  {
-    "Id": 0,
-    "Brand": 2,
-    "Model": "Corsa",
-    "DoorsQuantity": 5,
-    "Color": "White",
-    "Transmission": 0
-  },
-  {
-    "Id": 1,
-    "Brand": 5,
-    "Model": "Passat",
-    "DoorsQuantity": 4,
-    "Color": "Black",
-    "Transmission": 1
-  }
-]
